fix licence in converter_tools, change folder structure, break tests appart
